+++ resources/templates/ts/retained.ts
@@ -13,7 +13,7 @@ __disposables.push(
     if (editor) {
       vscode.window.showInformationMessage(
         `Active editor: ${vscode.workspace.asRelativePath(editor.document.uri)}. ` +
-        `Macro ${__runId} must be manually stopped for this listener to be disposed.`,
+          `Macro ${__runId} must be manually stopped for this listener to be disposed.`,
       );
     }
   }),